moved modal into button component
