+++ ClientApp/src/components/admin/overview/ConnectButton.tsx
@@ -1,20 +1,35 @@
-import { Button } from "@material-ui/core";
-import React from "react";
+import React, {useState} from "react";
 import useStyles from "./Styles";
 import { SelectedConnectionType } from "../../../common/components/Types";
+import { Grid, Typography, Modal, Box, Button} from "@material-ui/core";
+
 
 type ConnectButtonProps = {
   selectedConnection: SelectedConnectionType;
   connectGiverRecipient: () => void;
-  confirmModal: () => void; 
+};
+
+const style = {
+  position: 'absolute' as 'absolute',
+  top: '50%',
+  left: '50%',
+  transform: 'translate(-50%, -50%)',
+  width: 400,
+  bgcolor: "#e6e6e6",
+  border: '2px solid #000',
+  boxShadow: 24,
+  mx: 5, 
+  p: 5,
 };
 
 const ConnectButton: React.FC<ConnectButtonProps> = ({
   selectedConnection,
   connectGiverRecipient,
-  confirmModal,
 }) => {
   const classes = useStyles();
+  const [open, setOpen] = useState(false);
+  const openModal = () => setOpen(true);
+  const closeModal = () => setOpen(false);
   const displayButton = () => {
     if (
       selectedConnection.giver != undefined &&
@@ -49,14 +64,32 @@ const ConnectButton: React.FC<ConnectButtonProps> = ({
       selectedConnection.recipient != undefined
     ) {
       return (
+        <div>
         <Button
-          onClick={confirmModal}
+          onClick={() => {openModal();}}
           className={classes.finishedButton}
         >
           Koble sammen <br />
           {selectedConnection.giver?.fullName} med familienummer{" "}
           {selectedConnection.recipient?.familyId}
         </Button>
+        <Modal open={open}>
+        <Box sx={style}>
+          <Typography variant="h6" component="h2">
+            Bekreft at du vil koble {selectedConnection.giver?.fullName} sammen med familienummer {""}
+            {selectedConnection.recipient?.familyId}
+            </Typography>
+            <Grid container direction="row" spacing={8}>
+              <Grid item>
+                <Button color="primary" variant="contained" onClick={() => {connectGiverRecipient();closeModal();}}>Ja</Button>
+              </Grid>
+              <Grid item >
+                <Button color="inherit" variant="contained" onClick={closeModal}>Nei</Button>
+              </Grid>
+            </Grid>
+        </Box>
+        </Modal>
+        </div>
       );
     } else {
       return <></>;
