+++ ClientApp/src/components/admin/overview/OverviewMacro.tsx
@@ -1,4 +1,4 @@
-import { Container, Grid, Snackbar,Typography, Modal, Box, Button} from "@material-ui/core";
+import { Container, Grid, Snackbar,Typography} from "@material-ui/core";
 import React, {
   useCallback,
   useEffect,
@@ -24,28 +24,13 @@ interface IOverviewMacro {
   accessToken: string;
 }
 
-const style = {
-  position: 'absolute' as 'absolute',
-  top: '50%',
-  left: '50%',
-  transform: 'translate(-50%, -50%)',
-  width: 400,
-  bgcolor: "#e6e6e6",
-  border: '2px solid #000',
-  boxShadow: 24,
-  mx: 5, 
-  p: 5,
-};
   
 const OverviewMacro: React.FC<IOverviewMacro> = ({ accessToken, location }) => {
   const [selectedConnection, setSelectedConnection] =
     useState<SelectedConnectionType>(initState);
   const [giverData, setGiverData] = useState<GiverType[] | []>([]);
   const [recipientData, setRecipientData] = useState<RecipientType[] | []>([]);
-  const apiservice = new ApiService(accessToken);
-  const [open, setOpen] = useState(false);
-  const openModal = () => setOpen(true);
-  const closeModal = () => setOpen(false);
+  const apiservice = new ApiService(accessToken); 
 
   async function fetchGivers() {
     await apiservice
@@ -164,26 +149,9 @@ const OverviewMacro: React.FC<IOverviewMacro> = ({ accessToken, location }) => {
             />
           </Grid>
         </Grid>
-        <Modal open={open}>
-          <Box sx={style}>
-            <Typography variant="h6" component="h2">
-              Bekreft at du vil koble {selectedConnection.giver?.fullName} sammen med familienummer {""}
-              {selectedConnection.recipient?.familyId}
-              </Typography>
-              <Grid container direction="row" spacing={8}>
-                <Grid item>
-                  <Button color="primary" variant="contained" onClick={() => {connectGiverRecipient();closeModal();}}>Ja</Button>
-                </Grid>
-                <Grid item >
-                  <Button color="inherit" variant="contained" onClick={closeModal}>Nei</Button>
-                </Grid>
-              </Grid>
-          </Box>
-          </Modal>
         <ConnectButton
           selectedConnection={selectedConnection}
           connectGiverRecipient={connectGiverRecipient}
-          confirmModal={openModal}
         />
       </Container>
     </>
